+++ packages/engine/Source/Scene/PointCloudEyeDomeLighting.js
@@ -129,8 +129,7 @@ function getECShaderProgram(context, shaderProgram) {
     fs.sources.splice(
       0,
       0,
-      `layout (location = 0) out vec4 out_FragData_0;
-       layout (location = 1) out vec4 out_FragData_1;`
+      `layout (location = 0) out vec4 out_FragData_0;\nlayout (location = 1) out vec4 out_FragData_1;`
     );
 
     fs.sources = fs.sources.map(function (source) {